Remove support for Netscape cert extensions
Also added some explanatory comments.
Moved the to/from string<>ASN1_IA5 functions to another file.
