+++ crypto/include/internal/x509_int.h
@@ -171,7 +171,6 @@ struct x509_st {
     uint32_t ex_flags;
     uint32_t ex_kusage;
     uint32_t ex_xkusage;
-    uint32_t ex_nscert;
     ASN1_OCTET_STRING *skid;
     AUTHORITY_KEYID *akid;
     X509_POLICY_CACHE *policy_cache;
