+++ crypto/x509/standard_exts.h
@@ -13,14 +13,6 @@
  */
 
 static const X509V3_EXT_METHOD *standard_exts[] = {
-    &v3_nscert,
-    &v3_ns_ia5_list[0],
-    &v3_ns_ia5_list[1],
-    &v3_ns_ia5_list[2],
-    &v3_ns_ia5_list[3],
-    &v3_ns_ia5_list[4],
-    &v3_ns_ia5_list[5],
-    &v3_ns_ia5_list[6],
     &v3_skey_id,
     &v3_key_usage,
     &v3_pkey_usage_period,
