+++ crypto/x509/x_x509.c
@@ -63,7 +63,6 @@ static int x509_cb(int operation, ASN1_VALUE **pval, const ASN1_ITEM *it,
         ret->ex_cached = 0;
         ret->ex_kusage = 0;
         ret->ex_xkusage = 0;
-        ret->ex_nscert = 0;
         ret->ex_flags = 0;
         ret->ex_pathlen = -1;
         ret->ex_pcpathlen = -1;
